refactor: split `Program.cs`
